third_party/WebKit/LayoutTests/http/tests/devtools/compiler-source-mapping-debug.html

<html>
<head>
<script src="../inspector/inspector-test.js"></script>
<script src="../inspector/debugger-test.js"></script>
<script src="../inspector/console-test.js"></script>
<script src="../inspector/network-test.js"></script>
<script src="resources/compiled.js"></script>

<script>

function clickButton()
{
    document.getElementById('test').click();
}

function installScriptWithPoorSourceMap()
{
    var script = document.createElement("script");
    script.setAttribute("src", "resources/compiled-with-wrong-source-map-url.js");
    document.head.appendChild(script);
}

function test() {
  SourcesTestRunner.runDebuggerTestSuite([function testSetBreakpoint(next) {
    SourcesTestRunner.showScriptSource('source1.js', didShowSource);

    function didShowSource(sourceFrame) {
      TestRunner.addResult('Script source was shown.');
      SourcesTestRunner.setBreakpoint(sourceFrame, 14, '', true);
      SourcesTestRunner.waitUntilPaused(paused);
      TestRunner.evaluateInPage('setTimeout(clickButton, 0)');
    }

    function paused(callFrames) {
      SourcesTestRunner.captureStackTrace(callFrames);
      SourcesTestRunner.resumeExecution(next);
    }
  }]);
};

</script>

</head>

<body onload="runTest()">
<p>Tests installing compiler source map in scripts panel.</p>
</body>
</html>
